Move UserPublicProfileChip to a subfolder User
